fix(stella-cli): unbreak PR #2465 — split memory/tests.rs and repair the merged-clean call sites
Two independent failures on this branch, neither of them the prompt change
this PR is about.

**The file-size ratchet.** `crates/stella-cli/src/memory/tests.rs` reached
1503 lines — 1489 on `main` plus this PR's own 14 — and it carries no
baseline entry, so the guard is right to refuse it. AGENTS.md's answer to a
first crossing is a split and never a baseline row, so the reflection
lesson-mining seam moves to `memory/tests/reflection_mining.rs`: everything
from a model's reflection response to a stored lesson or self-review, plus
the three fixtures (`lessons_of`, `lessons_with`,
`reflection_execution_ids`) that only it reads. 763 and 764 lines.

Moved verbatim, and checked rather than asserted: the set of
`#[test]`/`#[tokio::test]` names under `crates/stella-cli/src/memory/` is
byte-identical before and after — 242 either side, no deletions, no
additions — which is the same question `scripts/check-deleted-tests.sh`
asks. Nothing to name in the PR description.

**A merge that compiled nothing.** `main` grew a seventh `posture:
ReflectionPosture` argument on `reflect_and_record`; #2482 landed
`memory/tests/trigger_recall.rs` afterwards holding two six-argument calls
written before that parameter existed. Neither side touched the other's
lines, so git reported no conflict and the branch has not built since. Both
call sites now pass `ReflectionPosture::default()`, which its own doc
comment defines as "exactly as before this was threaded" — the behaviour
those tests were written against, and what every other test in the tree
passes.

Verified: `cargo test -p stella-cli memory::` is 242 passed, 0 failed.

Refs #2465
